requirements.txt
streamlit
openpyxl
barcode
qrcode
aztec-code-generator
Pillow
